pyproject.toml
[project]
name = "signalgate"
version = "0.1.0"
description = "SignalGate: a personal signal gatekeeper (not an information system)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
]

[project.scripts]
signalgate = "signalgate.__main__:main"

[tool.setuptools]
packages = ["signalgate"]
